Don't delete store purchases when deleting a store
Store purchases would be kept undeleted when the related store is deleted.
However this change would force you to get the store association from a store purchase
with Paranoid filter "all" when you need to show a deleted store of a purchase.

Introduced new Model_Store_Purchase::store() method for just that.

This is not breaking BC unless you need the store association working
even for deleted stores.