Iterate arrays when freezing to avoid reference issues
When iterating an association and perform an action
which could access the parent model and then the same association
it is accessed by reference and nesting `foreach` loops
is causing the internal pointer of the Iterator to move
so the outer loop iteration is executed only once.

This may cause in multi-brand-purchase scenario,
only some of purchase items to be frozen
depending on the actions performed by the freeze.